requirements.txt
numpy~=1.23.3
requests~=2.28.1
scipy~=1.9.3
click~=8.1.3
pillow~=9.2.0
pyspng~=0.1.1
psutil~=5.9.0
tqdm~=4.64.1